move LexiCommand to LegacyLexiCommand for pre-prep
