+++ src/legacy-commands/uwus.ts
@@ -19,19 +19,19 @@
 import type { Message } from 'discord.js';
 
 import type CommandConfig from '#lib/interfaces/commands/LexiCommandConfig';
-import LexiCommand from '#lib/structures/LexiCommand';
+import LegacyLexiCommand from '#lib/structures/LegacyLexiCommand';
 import { formatRow } from '#lib/utils/formatTrackerRow';
 
 // The number of users that should be displayed on the leaderboard at a
 // given time.
 const CUTOFF = 10;
 
-export default class OwosCommand extends LexiCommand {
+export default class UwusCommand extends LegacyLexiCommand {
     public getConfig(): CommandConfig {
         return {
-            name: 'owos',
+            name: 'uwus',
             category: 'utility',
-            description: 'Shows a leaderboard of the messages containing "owo"!',
+            description: 'Shows a leaderboard of the messages containing "uwu"!',
             usage: '',
             enabled: true,
             aliases: [],
@@ -43,16 +43,16 @@ export default class OwosCommand extends LexiCommand {
             restrict: false
         };
     }
+
     public async run(message: Message): Promise<void> {
         const { client } = this;
-
-        const sorted = client.emoteCounterTrackers.map((values, id) => [id, values.owos] as const).sort((a, b) => b[1] - a[1]);
+        const sorted = client.emoteCounterTrackers.map((values, id) => [id, values.uwus] as const).sort((a, b) => b[1] - a[1]);
 
         // Whether or not the user of the command has placed in the top of the
         // leaderboard.
         let placed = false;
 
-        const buf = ['```adoc\n===== OWO LEADERBOARD ====='];
+        const buf = ['```adoc\n===== UWU LEADERBOARD ====='];
 
         for (let i = 0; i < sorted.length && i < CUTOFF; i++) {
             const [id] = sorted[i];
@@ -60,12 +60,12 @@ export default class OwosCommand extends LexiCommand {
 
             // eslint-disable-next-line no-await-in-loop
             const user = await client.users.fetch(id).catch(() => null);
-            buf.push(formatRow('owos', i, user, client));
+            buf.push(formatRow('uwus', i, user, client));
         }
 
         if (!placed) {
             const index = sorted.findIndex(([id]) => id === message.author.id);
-            if (index !== -1) buf.push('...', formatRow('owos', index, message.author, client));
+            if (index !== -1) buf.push('...', formatRow('uwus', index, message.author, client));
         }
 
         buf.push('```');
