+++ src/lib/structures/LegacyLexiCommand.ts
@@ -22,7 +22,7 @@ import type CommandConfig from '#lib/interfaces/commands/LexiCommandConfig';
 import type LexiLogger from '#lib/interfaces/LexiLogger';
 import type LexiClient from '#lib/structures/LexiClient';
 
-export default abstract class LexiCommand {
+export default abstract class LegacyLexiCommand {
     protected client: LexiClient;
     protected log: LexiLogger;
